WIP: refactoring hocuspocus and moved listen function to server instead
