+++ packages/server/src/index.ts
@@ -5,4 +5,5 @@ export * from './Hocuspocus.js'
 export * from './IncomingMessage.js'
 export * from './MessageReceiver.js'
 export * from './OutgoingMessage.js'
+export * from './Server.js'
 export * from './types.js'
