+++ playground/backend/.gitignore
@@ -1,2 +1,3 @@
 /database
 /dashboard
+/db.sqlite
