+++ playground/backend/src/express.ts
@@ -1,9 +1,9 @@
 import { Logger } from '@hocuspocus/extension-logger'
-import { Server } from '@hocuspocus/server'
+import { Hocuspocus } from '@hocuspocus/server'
 import express from 'express'
 import expressWebsockets from 'express-ws'
 
-const server = Server.configure({
+const hocuspocus = new Hocuspocus({
   extensions: [
     new Logger(),
   ],
@@ -17,7 +17,7 @@ app.get('/', (request, response) => {
 
 app.ws('/:documentName', (websocket, request: any) => {
   const context = { user_id: 1234 }
-  server.handleConnection(websocket, request, context)
+  hocuspocus.handleConnection(websocket, request, context)
 })
 
 app.listen(1234, () => console.log('Listening on http://127.0.0.1:1234…'))
